+++ src/db/models/Role.js
@@ -1,4 +1,0 @@
-const mongoose = require("mongoose");
-const roleSchema = require("../schemas/Role");
-const roleModel = mongoose.model("roles", roleSchema);
-module.exports = roleModel;