+++ tests/e2e/game/index.js
@@ -1,0 +1,6 @@
+const { describe } = require("mocha");
+
+describe("E2E - Game tests", () => {
+    require("./game-creation.test");
+    require("./simple-game.test");
+});